First stab at default user options file.
The programs look in user's home directory for file .PBSUtil.xml
(i.e., ~/.PBSUtil.xml) and if present, grabs the user defaults
that are in there. So far only user, host, email, and dir are
supported. An example config file is:

<defaults>
  <user>joe</user>
  <host>nexus.westgrid.ca</host>
  <email>joe1000@yahoo.com</email>
  <dir>/scratch/gumby</dir>
</defaults>

The logic used to calculate such variables is:

if variable is passed on the commandline:
   use that
else:
   if it is in ~/.PBSUtil.xml:
      use that instead
   else:
      try to figure out a default (like getting USER from env), or
      calculating email using the 'user' value, etc.